feat(Lesson10): Preparing Files for class
