+++ src/components/Button/Button.jsx
@@ -1,10 +1,14 @@
 import styles from "./Button.module.css";
 
 const Button = ({ isSelected, name, onClick, children }) => {
-  const buttonStyles = `${styles.button} ${isSelected ? styles.selected : ""}`;
+  const buttonStyles = [styles.button];
+
+  if (isSelected) {
+    buttonStyles.push(styles.selected);
+  }
 
   return (
-    <button className={buttonStyles} onClick={() => onClick(name)}>
+    <button className={buttonStyles.join(" ")} onClick={() => onClick(name)}>
       {children}
     </button>
   );
