+++ src/components/Button/Button.module.css
@@ -14,6 +14,10 @@
   &:hover {
     background-color: #0056b3;
   }
+
+  &:focus {
+    outline: none !important;
+  }
 }
 
 .selected {
