+++ src/components/Cart/Cart.module.css
@@ -1,0 +1,6 @@
+.cart {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 16px 0;
+}
