+++ src/exercises/lesson10/lesson10.jsx
@@ -1,0 +1,5 @@
+const lesson10 = () => {
+  return <div>lesson10</div>;
+};
+
+export default lesson10;
