+++ src/views/RestaurantView.jsx
@@ -1,9 +1,10 @@
-import "./RestaurantView.css";
-
 import { useState } from "react";
-import Cart from "../components/Cart.jsx";
-import MenuItem from "../components/MenuItem.jsx";
+import Cart from "../components/Cart/Cart.jsx";
+import MenuItem from "../components/MenuItem/MenuItem.jsx";
 import dishes from "../mocks/dishes.json";
+
+import styles from "./RestaurantView.module.css";
+import NavBar from "../components/NavBar/NavBar.jsx";
 
 const RestaurantView = () => {
   const [selectedItems, setSelectedItems] = useState([]);
@@ -22,23 +23,28 @@ const RestaurantView = () => {
 
   return (
     <>
-      <h1 style={{ textAlign: "center" }}>ReDI React Restaurant</h1>
-      <Cart
-        selectedItems={selectedItems}
-        onClear={() => setSelectedItems([])}
-      />
-      <div className="menu">
-        {dishes.map((dish) => (
-          <MenuItem
-            key={dish.name}
-            name={dish.name}
-            image={dish.image}
-            onClick={handleMenuClick}
-            isSelected={selectedItems.includes(dish.name)}
-          />
-        ))}
+      <NavBar>
+        <h1>ReDI React Restaurant</h1>
+      </NavBar>
+
+      <div className={styles.restaurantWrapper}>
+        <Cart
+          selectedItems={selectedItems}
+          onClear={() => setSelectedItems([])}
+        />
+        <div className={styles.menu}>
+          {dishes.map((dish) => (
+            <MenuItem
+              key={dish.name}
+              name={dish.name}
+              image={dish.image}
+              onClick={handleMenuClick}
+              isSelected={selectedItems.includes(dish.name)}
+            />
+          ))}
+        </div>
+        {selectedItems.length >= 3 && <h4>🎉 You unlocked a 10% discount!</h4>}
       </div>
-      {selectedItems.length >= 3 && <h4>🎉 You unlocked a 10% discount!</h4>}
     </>
   );
 };
